[#175][#140] Ensure we return Floats, not Fixnum or Rational; fixes segfaults with mathn

+++ lib/simplecov/file_list.rb
@@ -31,14 +31,16 @@ class SimpleCov::FileList < Array
   end
 
   # Computes the coverage based upon lines covered and lines missed
+  # @return [Float]
   def covered_percent
     return 100.0 if empty? or lines_of_code == 0
-    covered_lines * 100.0 / lines_of_code
+    Float(covered_lines * 100.0 / lines_of_code)
   end
 
   # Computes the strength (hits / line) based upon lines covered and lines missed
+  # @return [Float]
   def covered_strength
-    return 0 if empty? or lines_of_code == 0
-    map {|f| f.covered_strength * f.lines_of_code }.inject(&:+) / lines_of_code
+    return 0.0 if empty? or lines_of_code == 0
+    Float(map {|f| f.covered_strength * f.lines_of_code }.inject(&:+) / lines_of_code)
   end
 end
